move no kinematics functions
