+++ app/src/utilities/specification_mg400.py
@@ -1,0 +1,17 @@
+"""specification of mg400"""
+import numpy as np
+
+J1_MIN = -160
+J1_MAX = 160
+J2_MIN = -25
+J2_MAX = 85
+J3_MIN = -25
+J3_MAX = 105
+J3_1_MIN = -60
+J3_1_MAX = 60
+
+
+LINK1 = np.array([43, 0, 0])
+LINK2 = np.array([0, 0, 175])
+LINK3 = np.array([175, 0, 0])
+LINK4 = np.array([66, 0, -57])
